requirements.txt
numpy~=1.22.3
opencv-python~=4.5.3.56
torch>=1.7
torchvision~=0.12.0
loguru~=0.6.0
scikit-image~=0.18.2
tqdm~=4.64.0
Pillow~=9.1.0
thop~=0.0.31.post2005241907
ninja~=1.10.2.3
tabulate~=0.8.9
tensorboard~=2.8.0
pyyaml~=6.0
pycocotools~=2.0.2
flask~=2.1.2
timm~=0.5.4
requests~=2.27.1

# onnx==1.8.1
# onnxruntime==1.8.0
# onnx-simplifier==0.3.5

tensorrt~=0.0.1
xmltodict~=0.12.0
imgviz~=1.5.0
labelme~=5.0.1
beautifulsoup4~=4.11.0
soupsieve~=2.3.2
defusedxml~=0.7.1
cffi~=1.15.0
PyQt5~=5.15.6
ipython~=8.2.0
pip~=21.1.2
wheel~=0.36.2
cryptography~=37.0.1
tornado~=6.1
rsa~=4.8
pyasn1~=0.4.8
pyzmq~=22.3.0
colorama~=0.4.4
paramiko~=2.7.2
six~=1.16.0
gast~=0.5.3
h5py~=3.6.0
parso~=0.8.3
jedi~=0.18.1
scipy~=1.8.0
matplotlib~=3.2.2
QtPy~=2.0.1
packaging~=21.3
tensorflow-gpu~=2.8.0
keras~=2.8.0
ipywidgets~=7.7.0
notebook~=6.4.10
attrs~=21.4.0
MarkupSafe~=2.1.1
Werkzeug~=2.1.1
Jinja2~=3.1.1
click~=8.1.3
itsdangerous~=2.1.2
pywin32~=303
wrapt~=1.14.0
Cython~=0.29.28
setuptools~=57.0.0
astunparse~=1.6.3
psutil~=5.9.0
urllib3~=1.26.9
webencodings~=0.5.1
bleach~=5.0.0
networkx~=2.7.1
Pygments~=2.11.2
Django~=4.0.4
sqlparse~=0.4.2
asgiref~=3.5.0
cachetools~=5.0.0
certifi~=2021.10.8
pyzlib~=0.1.16
debugpy~=1.6.0
traitlets~=5.1.1
decorator~=5.1.1
imageio~=2.16.1
tifffile~=2022.3.25
testpath~=0.6.0
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.5
natsort~=8.1.0
termcolor~=1.1.0
python-dateutil~=2.8.2
Markdown~=3.3.6
nbformat~=5.3.0
nbclient~=0.5.13
nbconvert~=6.4.5
fastjsonschema~=2.15.3
jsonschema~=4.4.0
Send2Trash~=1.8.0
terminado~=0.13.3
oauthlib~=3.2.0
bcrypt~=3.2.0
idna~=3.3
asttokens~=2.0.5
executing~=0.8.3
fonttools~=4.31.2
ipykernel~=6.12.1
pandocfilters~=1.5.0
mistune~=0.8.4
entrypoints~=0.4
pyparsing~=3.0.7
pycparser~=2.21
qtconsole~=5.3.0
pyrsistent~=0.18.1
cycler~=0.11.0
kiwisolver~=1.4.2
flatbuffers~=2.0
zipp~=3.8.0
legacy~=0.1.6
pytz~=2022.1
pandas~=1.4.2
automium~=0.2.6
seaborn~=0.11.2
torchaudio~=0.11.0+cu115
PyMySQL~=1.0.2